requirements.txt
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu1186
numpy==2.2.6
pandas==2.3.3
matplotlib==3.7.0
tqdm==4.67.1
scikit-learn==1.7.2
opencv-python==4.12.0
pillow==12.0.0
triton==2.2.0
causal-conv1d==1.1.3
mamba-ssm==1.1.3
scipy==1.15.3